+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # dmptool-types CHANGELOG
 
+# v2.2.0
+- Added code to allow loading dmptool-types from github repo instead of npm package for pre-release testing.
+- Added to README instructions on how to load dmptool-types from github repo.
+- Added `showCommentField` and `comment` to most question and answers (except ones ticket said not to)
+
 # v2.1.0
 - Added script to fetch the latest RDA Common Standard JSON schema
 - Added logic to generate types from the RDA Common Standard JSON schema